fix(ui): TEKAA-14 — format tanggal panjang Indonesia di semua label terbit/diperbarui
Lanjutan PR #19 (timezone WIB): sisa label masih numeric pendek (11/8/2026).
Sekarang konsisten '11 Agustus 2026' via toLocaleDateString('id-ID', { month: 'long', timeZone: 'Asia/Jakarta' }) di:
- posts/[slug].astro (label Terbit + Diperbarui)
- PostCard, EditorsPicks, TrendingStrip, index.astro
- og/[slug].png.ts (label tanggal OG image)

Data mesin (JSON-LD, RSS, sitemap, llms.txt) tetap ISO — hanya tampilan visible.

+++ src/pages/og/[slug].png.ts
@@ -43,7 +43,7 @@ export async function GET({ params }: { params: { slug: string } }) {
         title: post.data.title,
         category: post.data.category,
         author: post.data.author,
-        publishedAt: new Date(post.data.publishedAt).toLocaleDateString('id-ID', { day: 'numeric', month: 'numeric', year: 'numeric', timeZone: 'Asia/Jakarta' }),
+        publishedAt: new Date(post.data.publishedAt).toLocaleDateString('id-ID', { day: 'numeric', month: 'long', year: 'numeric', timeZone: 'Asia/Jakarta' }),
       }),
       {
         // 16:9 sesuai standar OG TeknoPulse (standar-on-page-seo §6).
